pyproject.toml
[project]
name = "aoc-2025"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "pytest",
    "python-dotenv",
    "requests",
]

[tool.pytest.ini_options]
pythonpath = ["."]
